# contrat-api/openapi.yaml
openapi: 3.0.3
info:
  title: SenYummies API
  description: >
    API commune du projet SenYummies, developpe pour le restaurant Chez Ketchup.
    Consommee par les deux frontends : App Client (public, commande invite) et
    SenYummies Manager (interne, authentifie par role). Ce contrat est la source
    de verite entre les 3 equipes : toute modification passe par une revue du
    lead Backend avant merge.
  version: "0.1.0"
servers:
  - url: http://localhost:8080/api/v1
    description: Environnement de developpement local
tags:
  - name: Authentification
  - name: Produits
  - name: Commandes
  - name: Paiements
security: []
paths:
  /auth/login:
    post:
      tags: [Authentification]
      summary: Connexion du staff SenYummies Manager
      description: Seul le back-office authentifie ses utilisateurs. L'App Client ne l'appelle jamais.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion reussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /auth/first-login-password:
    post:
      tags: [Authentification]
      summary: Remplacer le mot de passe temporaire avant la premiere connexion
      description: Cette route publique n'emet un JWT qu'apres validation du mot de passe temporaire et enregistrement du nouveau mot de passe.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirstLoginPasswordRequest'
      responses:
        '200':
          description: Mot de passe remplace, session staff ouverte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Nouveau mot de passe invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Identifiant ou mot de passe temporaire incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /staff-users:
    get:
      tags: [Authentification]
      summary: Lister les comptes du personnel (Gerant uniquement)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Comptes staff sans donnees de mot de passe
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffUser'
        '403':
          description: Role Gerant requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags: [Authentification]
      summary: Creer un compte du personnel (Gerant uniquement)
      description: Genere un mot de passe temporaire retourne une seule fois. Le compte devra le remplacer avant d'acceder au Manager.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffUserCreateRequest'
      responses:
        '201':
          description: Compte cree avec identifiant et mot de passe temporaire a communiquer au staff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryPasswordResponse'
        '409':
          description: Identifiant deja utilise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /staff-users/{id}:
    patch:
      tags: [Authentification]
      summary: Modifier un compte du personnel (Gerant uniquement)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StaffUserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffUserUpdateRequest'
      responses:
        '200':
          description: Compte mis a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffUser'
        '403':
          description: Role Gerant requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Compte introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags: [Authentification]
      summary: Desactiver un compte du personnel (Gerant uniquement)
      description: Desactivation reversible, sans suppression de l'historique. Le dernier Gerant actif et le compte courant ne peuvent pas etre desactives.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StaffUserId'
      responses:
        '204':
          description: Compte desactive
        '409':
          description: Le dernier Gerant actif ou le compte courant ne peut pas etre desactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /staff-users/{id}/permanent:
    delete:
      tags: [Authentification]
      summary: Supprimer definitivement un compte du personnel desactive (Gerant uniquement)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StaffUserId'
      responses:
        '204':
          description: Compte supprime
        '409':
          description: Le compte doit d'abord etre desactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /staff-users/{id}/temporary-password:
    post:
      tags: [Authentification]
      summary: Regenerer un mot de passe temporaire (Gerant uniquement)
      description: Invalide le mot de passe precedent et force son remplacement avant le prochain acces.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StaffUserId'
      responses:
        '200':
          description: Mot de passe temporaire retourne une seule fois
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryPasswordResponse'
        '403':
          description: Role Gerant requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Compte introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /produits:
    get:
      tags: [Produits]
      summary: Liste du menu, visible sans authentification
      security: []
      responses:
        '200':
          description: Liste des produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produit'

  /commandes:
    post:
      tags: [Commandes]
      summary: Creation d'une commande par un client invite (sans compte)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandeCreation'
      responses:
        '201':
          description: Commande creee (statut initial CREEE, en attente de paiement)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commande'
        '400':
          description: Requete invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags: [Commandes]
      summary: Liste des commandes (staff uniquement)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des commandes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Commande'
        '401':
          description: Non authentifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /commandes/{id}:
    get:
      tags: [Commandes]
      summary: Suivi d'une commande par son identifiant (client invite)
      security: []
      parameters:
        - $ref: '#/components/parameters/CommandeId'
      responses:
        '200':
          description: Etat de la commande
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commande'
        '404':
          description: Commande introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /commandes/{id}/statut:
    patch:
      tags: [Commandes]
      summary: Mise a jour du statut par le staff (Cuisine, Livreur, Gerant, Manager selon le statut)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CommandeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [statut]
              properties:
                statut:
                  $ref: '#/components/schemas/StatutCommande'
      responses:
        '200':
          description: Commande mise a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commande'
        '403':
          description: Role insuffisant pour cette transition de statut
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /paiements/init:
    post:
      tags: [Paiements]
      summary: Initialisation du paiement d'une commande
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [commandeId]
              properties:
                commandeId:
                  type: string
      responses:
        '200':
          description: Reference et lien de paiement a suivre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaiementInit'

  /paiements/webhook:
    post:
      tags: [Paiements]
      summary: Callback du prestataire de paiement (PayDunya ou API directe, selon choix du PDG)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Format specifique au prestataire retenu (a preciser avec le PDG)
      responses:
        '200':
          description: Callback pris en compte

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CommandeId:
      name: id
      in: path
      required: true
      schema:
        type: string
    StaffUserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
          description: Le role MANAGER est unique et ne peut pas etre attribue via cette route.
        mustChangePassword:
          type: boolean
          description: Vrai quand la connexion utilise un mot de passe temporaire et qu'aucun JWT n'est emis.
    FirstLoginPasswordRequest:
      type: object
      required: [username, temporaryPassword, newPassword]
      properties:
        username:
          type: string
        temporaryPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
    StaffUserCreateRequest:
      type: object
      required: [firstName, lastName, address, cni, phone, username, role]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        cni:
          type: string
          pattern: '^[0-9]{14}$'
        phone:
          type: string
        username:
          type: string
          minLength: 3
        role:
          $ref: '#/components/schemas/Role'
          description: Le role MANAGER est reserve au compte unique du systeme.
    StaffUserUpdateRequest:
      type: object
      required: [username, role, active]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        cni:
          type: string
          pattern: '^[0-9]{14}$'
        phone:
          type: string
        username:
          type: string
          minLength: 3
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
    StaffUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        cni:
          type: string
        phone:
          type: string
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
        mustChangePassword:
          type: boolean
    TemporaryPasswordResponse:
      type: object
      required: [user, temporaryPassword]
      properties:
        user:
          $ref: '#/components/schemas/StaffUser'
        temporaryPassword:
          type: string
    Role:
      type: string
      enum: [CUISINE, GERANT, MANAGER, LIVREUR]
    Produit:
      type: object
      properties:
        id:
          type: string
        nom:
          type: string
        description:
          type: string
        prix:
          type: number
          format: double
        disponible:
          type: boolean
    LigneCommande:
      type: object
      required: [produitId, quantite]
      properties:
        produitId:
          type: string
        quantite:
          type: integer
          minimum: 1
    CommandeCreation:
      type: object
      required: [lignes, telephone, mode]
      properties:
        lignes:
          type: array
          items:
            $ref: '#/components/schemas/LigneCommande'
        telephone:
          type: string
        mode:
          type: string
          enum: [RETRAIT, LIVRAISON]
        adresse:
          type: string
          description: Requis uniquement si mode = LIVRAISON
    StatutCommande:
      type: string
      enum: [CREEE, PAYEE, EN_PREPARATION, PRETE, EN_LIVRAISON, LIVREE, RETIREE]
    Commande:
      type: object
      properties:
        id:
          type: string
        statut:
          $ref: '#/components/schemas/StatutCommande'
        lignes:
          type: array
          items:
            $ref: '#/components/schemas/LigneCommande'
        total:
          type: number
          format: double
        mode:
          type: string
          enum: [RETRAIT, LIVRAISON]
        creeLe:
          type: string
          format: date-time
    PaiementInit:
      type: object
      properties:
        paymentReference:
          type: string
        redirectUrl:
          type: string
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        field:
          type: string
          nullable: true